+++ thirdparty/sentry-native/src/backends/sentry_backend_crashpad.cpp
@@ -31,7 +31,11 @@ extern "C" {
 #include "client/crash_report_database.h"
 #include "client/crashpad_client.h"
 #include "client/crashpad_info.h"
+#include "client/prune_crash_reports.h"
 #include "client/settings.h"
+#if defined(_MSC_VER)
+#    include "util/win/termination_codes.h"
+#endif
 
 #if defined(__GNUC__)
 #    pragma GCC diagnostic pop
@@ -42,6 +46,7 @@ extern "C" {
 extern "C" {
 
 #ifdef SENTRY_PLATFORM_LINUX
+#    include <unistd.h>
 #    define SIGNAL_STACK_SIZE 65536
 static stack_t g_signal_stack;
 
@@ -121,44 +126,63 @@ sentry__crashpad_backend_flush_scope(
 #if defined(SENTRY_PLATFORM_LINUX) || defined(SENTRY_PLATFORM_WINDOWS)
 #    ifdef SENTRY_PLATFORM_WINDOWS
 static bool
-sentry__crashpad_handler(EXCEPTION_POINTERS *UNUSED(ExceptionInfo))
+sentry__crashpad_handler(EXCEPTION_POINTERS *ExceptionInfo)
 {
 #    else
 static bool
-sentry__crashpad_handler(int UNUSED(signum), siginfo_t *UNUSED(info),
-    ucontext_t *UNUSED(user_context))
+sentry__crashpad_handler(int signum, siginfo_t *info, ucontext_t *user_context)
 {
     sentry__page_allocator_enable();
     sentry__enter_signal_handler();
 #    endif
     SENTRY_DEBUG("flushing session and queue before crashpad handler");
 
-    SENTRY_WITH_OPTIONS (options) {
-        sentry__write_crash_marker(options);
+    bool should_dump = true;
+    sentry_value_t event = sentry_value_new_event();
 
-        sentry_value_t event = sentry_value_new_event();
-        if (options->before_send_func) {
+    SENTRY_WITH_OPTIONS (options) {
+
+        if (options->on_crash_func) {
+            sentry_ucontext_t uctx;
+#    ifdef SENTRY_PLATFORM_WINDOWS
+            uctx.exception_ptrs = *ExceptionInfo;
+#    else
+            uctx.signum = signum;
+            uctx.siginfo = info;
+            uctx.user_context = user_context;
+#    endif
+
+            SENTRY_TRACE("invoking `on_crash` hook");
+            event
+                = options->on_crash_func(&uctx, event, options->on_crash_data);
+        } else if (options->before_send_func) {
             SENTRY_TRACE("invoking `before_send` hook");
             event = options->before_send_func(
-                event, NULL, options->before_send_data);
+                event, nullptr, options->before_send_data);
         }
+        should_dump = !sentry_value_is_null(event);
         sentry_value_decref(event);
 
-        sentry__record_errors_on_current_session(1);
-        sentry_session_t *session = sentry__end_current_session_with_status(
-            SENTRY_SESSION_STATUS_CRASHED);
-        if (session) {
-            sentry_envelope_t *envelope = sentry__envelope_new();
-            sentry__envelope_add_session(envelope, session);
+        if (should_dump) {
+            sentry__write_crash_marker(options);
 
-            // capture the envelope with the disk transport
-            sentry_transport_t *disk_transport
-                = sentry_new_disk_transport(options->run);
-            sentry__capture_envelope(disk_transport, envelope);
-            sentry__transport_dump_queue(disk_transport, options->run);
-            sentry_transport_free(disk_transport);
+            sentry__record_errors_on_current_session(1);
+            sentry_session_t *session = sentry__end_current_session_with_status(
+                SENTRY_SESSION_STATUS_CRASHED);
+            if (session) {
+                sentry_envelope_t *envelope = sentry__envelope_new();
+                sentry__envelope_add_session(envelope, session);
+
+                // capture the envelope with the disk transport
+                sentry_transport_t *disk_transport
+                    = sentry_new_disk_transport(options->run);
+                sentry__capture_envelope(disk_transport, envelope);
+                sentry__transport_dump_queue(disk_transport, options->run);
+                sentry_transport_free(disk_transport);
+            }
+        } else {
+            SENTRY_TRACE("event was discarded");
         }
-
         sentry__transport_dump_queue(options->transport, options->run);
     }
 
@@ -166,6 +190,41 @@ sentry__crashpad_handler(int UNUSED(signum), siginfo_t *UNUSED(info),
 #    ifndef SENTRY_PLATFORM_WINDOWS
     sentry__leave_signal_handler();
 #    endif
+
+    // If we __don't__ want a minidump produced by crashpad we need to either
+    // exit or longjmp at this point. The crashpad client handler which calls
+    // back here (SetFirstChanceExceptionHandler) does the same if the
+    // application is not shutdown via the crashpad_handler process.
+    //
+    // If we would return `true` here without changing any of the global signal-
+    // handling state or rectifying the cause of the signal, this would turn
+    // into a signal-handler/exception-filter loop, because some
+    // signals/exceptions (like SIGSEGV) are unrecoverable.
+    //
+    // Ideally the SetFirstChanceExceptionHandler would accept more than a
+    // boolean to differentiate between:
+    //
+    // * we accept our fate and want a minidump (currently returning `false`)
+    // * we accept our fate and don't want a minidump (currently not available)
+    // * we rectified the situation, so crashpads signal-handler can simply
+    //   return, thereby letting the not-rectified signal-cause trigger a
+    //   signal-handler/exception-filter again, which probably leads to us
+    //   (currently returning `true`)
+    //
+    // TODO(supervacuus):
+    // * we need integration tests for more signal/exception types not only
+    //   for unmapped memory access (which is the current crash in example.c).
+    // * we should adapt the SetFirstChanceExceptionHandler interface in
+    // crashpad
+    if (!should_dump) {
+#    ifdef SENTRY_PLATFORM_WINDOWS
+        TerminateProcess(GetCurrentProcess(),
+            crashpad::TerminationCodes::kTerminationCodeCrashNoDump);
+#    else
+        _exit(EXIT_FAILURE);
+#    endif
+    }
+
     // we did not "handle" the signal, so crashpad should do that.
     return false;
 }
@@ -408,14 +467,6 @@ sentry__crashpad_backend_last_crash(sentry_backend_t *backend)
     uint64_t crash_time = 0;
 
     std::vector<crashpad::CrashReportDatabase::Report> reports;
-    if (data->db->GetPendingReports(&reports)
-        == crashpad::CrashReportDatabase::kNoError) {
-        for (const crashpad::CrashReportDatabase::Report &report : reports) {
-            report_crash_time(&crash_time, report);
-        }
-    }
-
-    reports.clear();
     if (data->db->GetCompletedReports(&reports)
         == crashpad::CrashReportDatabase::kNoError) {
         for (const crashpad::CrashReportDatabase::Report &report : reports) {
@@ -424,6 +475,22 @@ sentry__crashpad_backend_last_crash(sentry_backend_t *backend)
     }
 
     return crash_time;
+}
+
+static void
+sentry__crashpad_backend_prune_database(sentry_backend_t *backend)
+{
+    crashpad_state_t *data = (crashpad_state_t *)backend->data;
+
+    // We want to eagerly clean up reports older than 2 days, and limit the
+    // complete database to a maximum of 8M. That might still be a lot for
+    // an embedded use-case, but minidumps on desktop can sometimes be quite
+    // large.
+    data->db->CleanDatabase(60 * 60 * 24 * 2);
+    crashpad::BinaryPruneCondition condition(crashpad::BinaryPruneCondition::OR,
+        new crashpad::DatabaseSizePruneCondition(1024 * 8),
+        new crashpad::AgePruneCondition(2));
+    crashpad::PruneCrashReportDatabase(data->db, &condition);
 }
 
 sentry_backend_t *
@@ -451,6 +518,7 @@ sentry__backend_new(void)
     backend->user_consent_changed_func
         = sentry__crashpad_backend_user_consent_changed;
     backend->get_last_crash_func = sentry__crashpad_backend_last_crash;
+    backend->prune_database_func = sentry__crashpad_backend_prune_database;
     backend->data = data;
     backend->can_capture_after_shutdown = true;