+++ thirdparty/sentry-native/src/sentry_core.c
@@ -16,6 +16,7 @@
 #include "sentry_session.h"
 #include "sentry_string.h"
 #include "sentry_sync.h"
+#include "sentry_tracing.h"
 #include "sentry_transport.h"
 #include "sentry_value.h"
 
@@ -23,12 +24,11 @@
 #    include "integrations/sentry_integration_qt.h"
 #endif
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
-#    include "sentry_tracing.h"
-#endif
-
 static sentry_options_t *g_options = NULL;
 static sentry_mutex_t g_options_lock = SENTRY__MUTEX_INIT;
+
+/// see sentry_get_crashed_last_run() for the possible values
+static int g_last_crash = -1;
 
 const sentry_options_t *
 sentry__options_getref(void)
@@ -161,6 +161,7 @@ sentry_init(sentry_options_t *options)
         last_crash = backend->get_last_crash_func(backend);
     }
 
+    g_last_crash = sentry__has_crash_marker(options);
     g_options = options;
 
     // *after* setting the global options, trigger a scope and consent flush,
@@ -181,7 +182,11 @@ sentry_init(sentry_options_t *options)
 
     // after initializing the transport, we will submit all the unsent envelopes
     // and handle remaining sessions.
+    SENTRY_TRACE("processing and pruning old runs");
     sentry__process_old_runs(options, last_crash);
+    if (backend && backend->prune_database_func) {
+        backend->prune_database_func(backend);
+    }
 
     if (options->auto_session_tracking) {
         sentry_start_session();
@@ -367,27 +372,29 @@ event_is_considered_error(sentry_value_t event)
     return false;
 }
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
 bool
 sentry__event_is_transaction(sentry_value_t event)
 {
     sentry_value_t event_type = sentry_value_get_by_key(event, "type");
     return sentry__string_eq("transaction", sentry_value_as_string(event_type));
 }
-#endif
 
 sentry_uuid_t
 sentry_capture_event(sentry_value_t event)
 {
-#ifdef SENTRY_PERFORMANCE_MONITORING
     if (sentry__event_is_transaction(event)) {
         return sentry_uuid_nil();
     } else {
         return sentry__capture_event(event);
     }
-#else
-    return sentry__capture_event(event);
-#endif
+}
+
+bool
+sentry__roll_dice(double probability)
+{
+    uint64_t rnd;
+    return probability >= 1.0 || sentry__getrandom(&rnd, sizeof(rnd))
+        || ((double)rnd / (double)UINT64_MAX) <= probability;
 }
 
 sentry_uuid_t
@@ -401,15 +408,11 @@ sentry__capture_event(sentry_value_t event)
     SENTRY_WITH_OPTIONS (options) {
         was_captured = true;
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
         if (sentry__event_is_transaction(event)) {
             envelope = sentry__prepare_transaction(options, event, &event_id);
         } else {
-            envelope = sentry__prepare_event(options, event, &event_id);
+            envelope = sentry__prepare_event(options, event, &event_id, true);
         }
-#else
-        envelope = sentry__prepare_event(options, event, &event_id);
-#endif
         if (envelope) {
             if (options->session) {
                 sentry_options_t *mut_options = sentry__options_lock();
@@ -421,8 +424,15 @@ sentry__capture_event(sentry_value_t event)
                 mut_options->session->init = false;
                 sentry__options_unlock();
             }
-            sentry__capture_envelope(options->transport, envelope);
-            was_sent = true;
+
+            bool should_skip = !sentry__roll_dice(options->sample_rate);
+            if (should_skip) {
+                SENTRY_DEBUG("throwing away event due to sample rate");
+                sentry_envelope_free(envelope);
+            } else {
+                sentry__capture_envelope(options->transport, envelope);
+                was_sent = true;
+            }
         }
     }
     if (!was_captured) {
@@ -431,15 +441,6 @@ sentry__capture_event(sentry_value_t event)
     return was_sent ? event_id : sentry_uuid_nil();
 }
 
-bool
-sentry__roll_dice(double probability)
-{
-    uint64_t rnd;
-    return probability >= 1.0 || sentry__getrandom(&rnd, sizeof(rnd))
-        || ((double)rnd / (double)UINT64_MAX) <= probability;
-}
-
-#ifdef SENTRY_PERFORMANCE_MONITORING
 bool
 sentry__should_send_transaction(sentry_value_t tx_cxt)
 {
@@ -456,22 +457,15 @@ sentry__should_send_transaction(sentry_value_t tx_cxt)
     }
     return send;
 }
-#endif
 
 sentry_envelope_t *
 sentry__prepare_event(const sentry_options_t *options, sentry_value_t event,
-    sentry_uuid_t *event_id)
+    sentry_uuid_t *event_id, bool invoke_before_send)
 {
     sentry_envelope_t *envelope = NULL;
 
     if (event_is_considered_error(event)) {
         sentry__record_errors_on_current_session(1);
-    }
-
-    bool should_skip = !sentry__roll_dice(options->sample_rate);
-    if (should_skip) {
-        SENTRY_DEBUG("throwing away event due to sample rate");
-        goto fail;
     }
 
     SENTRY_WITH_SCOPE (scope) {
@@ -483,7 +477,7 @@ sentry__prepare_event(const sentry_options_t *options, sentry_value_t event,
         sentry__scope_apply_to_event(scope, options, event, mode);
     }
 
-    if (options->before_send_func) {
+    if (options->before_send_func && invoke_before_send) {
         SENTRY_TRACE("invoking `before_send` hook");
         event
             = options->before_send_func(event, NULL, options->before_send_data);
@@ -524,7 +518,6 @@ fail:
     return NULL;
 }
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
 sentry_envelope_t *
 sentry__prepare_transaction(const sentry_options_t *options,
     sentry_value_t transaction, sentry_uuid_t *event_id)
@@ -555,7 +548,6 @@ fail:
     sentry_value_decref(transaction);
     return NULL;
 }
-#endif
 
 void
 sentry_handle_exception(const sentry_ucontext_t *uctx)
@@ -723,11 +715,9 @@ sentry_set_transaction(const char *transaction)
         sentry_free(scope->transaction);
         scope->transaction = sentry__string_clone(transaction);
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
         if (scope->transaction_object) {
             sentry_transaction_set_name(scope->transaction_object, transaction);
         }
-#endif
     }
 }
 
@@ -739,7 +729,6 @@ sentry_set_level(sentry_level_t level)
     }
 }
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
 sentry_transaction_t *
 sentry_transaction_start(
     sentry_transaction_context_t *opaque_tx_cxt, sentry_value_t sampling_ctx)
@@ -1019,4 +1008,21 @@ fail:
     sentry__span_free(opaque_span);
     return;
 }
-#endif
+
+int
+sentry_get_crashed_last_run()
+{
+    return g_last_crash;
+}
+
+int
+sentry_clear_crashed_last_run()
+{
+    bool success = false;
+    sentry_options_t *options = sentry__options_lock();
+    if (options) {
+        success = sentry__clear_crash_marker(options);
+    }
+    sentry__options_unlock();
+    return success ? 0 : 1;
+}