+++ thirdparty/sentry-native/src/sentry_options.c
@@ -33,6 +33,9 @@ sentry_options_new(void)
     opts->release = sentry__string_clone(getenv("SENTRY_RELEASE"));
     opts->environment = sentry__string_clone(getenv("SENTRY_ENVIRONMENT"));
 #endif
+    if (!opts->environment) {
+        opts->environment = sentry__string_clone("production");
+    }
     opts->max_breadcrumbs = SENTRY_BREADCRUMBS_MAX;
     opts->user_consent = SENTRY_USER_CONSENT_UNKNOWN;
     opts->auto_session_tracking = true;
@@ -51,11 +54,8 @@ sentry_options_new(void)
     opts->sample_rate = 1.0;
     opts->refcount = 1;
     opts->shutdown_timeout = SENTRY_DEFAULT_SHUTDOWN_TIMEOUT;
-
-#ifdef SENTRY_PERFORMANCE_MONITORING
     opts->traces_sample_rate = 0.0;
     opts->max_spans = 0;
-#endif
 
     return opts;
 }
@@ -120,6 +120,14 @@ sentry_options_set_before_send(
 {
     opts->before_send_func = func;
     opts->before_send_data = data;
+}
+
+void
+sentry_options_set_on_crash(
+    sentry_options_t *opts, sentry_crash_function_t func, void *data)
+{
+    opts->on_crash_func = func;
+    opts->on_crash_data = data;
 }
 
 void
@@ -378,7 +386,6 @@ sentry_options_set_database_pathw(sentry_options_t *opts, const wchar_t *path)
 }
 #endif
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
 /**
  * Sets the maximum number of spans that can be attached to a
  * transaction.
@@ -429,4 +436,10 @@ sentry_options_get_traces_sample_rate(sentry_options_t *opts)
 {
     return opts->traces_sample_rate;
 }
-#endif
+
+void
+sentry_options_set_backend(sentry_options_t *opts, sentry_backend_t *backend)
+{
+    sentry__backend_free(opts->backend);
+    opts->backend = backend;
+}