+++ thirdparty/sentry-native/src/sentry_uuid.c
@@ -102,28 +102,26 @@ sentry_uuid_as_string(const sentry_uuid_t *uuid, char str[37])
 #undef B
 }
 
-#ifdef SENTRY_PERFORMANCE_MONITORING
 void
 sentry__internal_uuid_as_string(const sentry_uuid_t *uuid, char str[37])
 {
-#    define B(X) (unsigned char)uuid->bytes[X]
+#define B(X) (unsigned char)uuid->bytes[X]
     snprintf(str, 33,
         "%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%"
         "02hhx%02hhx%02hhx%02hhx%02hhx%02hhx",
         B(0), B(1), B(2), B(3), B(4), B(5), B(6), B(7), B(8), B(9), B(10),
         B(11), B(12), B(13), B(14), B(15));
-#    undef B
+#undef B
 }
 
 void
 sentry__span_uuid_as_string(const sentry_uuid_t *uuid, char str[17])
 {
-#    define B(X) (unsigned char)uuid->bytes[X]
+#define B(X) (unsigned char)uuid->bytes[X]
     snprintf(str, 17, "%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx%02hhx", B(0),
         B(1), B(2), B(3), B(4), B(5), B(6), B(7));
-#    undef B
+#undef B
 }
-#endif
 
 #ifdef SENTRY_PLATFORM_WINDOWS
 sentry_uuid_t