+++ thirdparty/sentry-native/tests/unit/test_uninit.c
@@ -1,5 +1,4 @@
 #include "sentry_testsupport.h"
-#include <sentry.h>
 
 SENTRY_TEST(uninitialized)
 {